+++ syntree.h
@@ -37,7 +37,7 @@ struct Syntree_node {
     struct Type_info ntype;
     union SE_kind se;
     union Information info;
-    int tmppos;
+    size_t tmppos;
 
     int nodeid;
 };
